test(style): migrate inline tests to integration tests
Move public API tests from src/style/ to tests/style_tests.rs:
- Properties tests (Style, Color, Size, Spacing, etc.)
- Computed style tests
- Theme tests (Theme, ThemeManager, Themes presets)
- Theme signal tests (use_theme, set_theme, toggle_theme)
- Transition tests (Easing, Transition, TransitionManager)
- Animation easing tests
- Error types tests (ErrorCode, Severity, SourceLocation, etc.)

Keep inline tests that use private APIs:
- parser.rs: parse(), parse_color(), parse_size(), etc.
- animation/mod.rs: Tween private fields (from, to, delay)
- transition.rs: parse_duration()
- error.rs: levenshtein_distance()

159 integration tests + 103 inline tests = 262 total style tests
