+++ src/style/theme_signal.rs
@@ -291,86 +291,4 @@ pub fn theme_to_css_variables(theme: &Theme) -> String {
     )
 }
 
-// ─────────────────────────────────────────────────────────────────────────────
-// Tests
-// ─────────────────────────────────────────────────────────────────────────────
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::style::Themes;
-
-    #[test]
-    fn test_use_theme() {
-        let theme = use_theme();
-        let current = theme.get();
-        // Default is dark theme
-        assert!(!current.name.is_empty());
-    }
-
-    #[test]
-    fn test_set_theme() {
-        let theme = use_theme();
-        set_theme(Themes::nord());
-        assert_eq!(theme.get().name, "Nord");
-    }
-
-    #[test]
-    fn test_toggle_theme() {
-        let theme = use_theme();
-
-        // Set to dark first
-        set_theme(Theme::dark());
-        assert!(theme.get().is_dark());
-
-        // Toggle to light
-        toggle_theme();
-        assert!(theme.get().is_light());
-
-        // Toggle back to dark
-        toggle_theme();
-        assert!(theme.get().is_dark());
-    }
-
-    #[test]
-    fn test_register_theme() {
-        register_theme("test-theme", Theme::custom("Test").build());
-        assert!(theme_ids().contains(&"test-theme".to_string()));
-    }
-
-    #[test]
-    fn test_set_theme_by_id() {
-        register_theme("id-test", Themes::monokai());
-        assert!(set_theme_by_id("id-test"));
-        assert_eq!(use_theme().get().name, "Monokai");
-    }
-
-    #[test]
-    fn test_set_theme_by_invalid_id() {
-        assert!(!set_theme_by_id("nonexistent-theme-id"));
-    }
-
-    #[test]
-    fn test_get_theme() {
-        register_theme("get-test", Themes::dracula());
-        let theme = get_theme("get-test");
-        assert!(theme.is_some());
-        assert_eq!(theme.unwrap().name, "Dracula");
-    }
-
-    #[test]
-    fn test_theme_to_css_variables() {
-        let css = theme_to_css_variables(&Themes::dracula());
-        assert!(css.contains("--theme-bg:"));
-        assert!(css.contains("--theme-primary:"));
-        assert!(css.contains("Dracula"));
-    }
-
-    #[test]
-    fn test_theme_ids() {
-        let ids = theme_ids();
-        // Should have default themes registered
-        assert!(ids.contains(&"dark".to_string()));
-        assert!(ids.contains(&"light".to_string()));
-    }
-}
+// Tests moved to tests/style_tests.rs
